refactor: split YourAccountView god-class into per-flow sub-views
Extracts LoginView, RegistrationView, PasswordResetView, ActivationView,
and a shared YourAccountIcons trait from the 536-line YourAccountView
god-class (backlog 1.15). YourAccountView becomes a thin delegating
facade implementing YourAccountViewInterface unchanged, so the module
call site (modules/YourAccount/index.php) requires no changes.

Method bodies moved verbatim (byte-identical) per Decision B grouping:
- LoginView: renderLoginPage
- RegistrationView: renderRegisterPage, renderRegistrationCompletePage,
  renderRegistrationErrorPage
- PasswordResetView: renderForgotPasswordPage, renderResetEmailSentPage,
  renderResetPasswordPage, renderPasswordResetSuccessPage,
  renderPasswordResetErrorPage
- ActivationView: renderActivationSuccessPage, renderActivationErrorPage
- YourAccountIcons trait: 7 shared SVG icon helpers

Safety net: new byte-identical characterization test
(YourAccountViewCharacterizationTest) captures pre-refactor fixture
snapshots of all 13 render variants and normalizes the non-deterministic
CSRF token before comparison. All 37 existing YourAccountView/Xss tests
still pass, full suite green, PHPStan clean.
